feat: add cutoff date picker and refactor cutoff dropdown and leftpanel
